[BE] fix: 여행지 정보 도메인 필드 타입 수정

+++ backend/service/trip-service/src/main/java/com/followfollowme/tripmarble/domainlayer/trip/domain/model/TripSpot.java
@@ -5,7 +5,7 @@ import lombok.Builder;
 @Builder
 public record TripSpot(
     long id,
-    long contentTypeId,
+    int contentTypeId,
     int contentId,
     String title,
     String tel,
